requirements.txt
# C++ Interface Analyzer - Dependencies

# Core dependencies
python-dotenv>=1.0.0

# LLM Integration
langchain>=0.1.0
langchain-ollama>=0.1.0
langchain-openai>=0.1.0

# Optional: For Phase 2 (AST parsing)
# libclang>=16.0.0
# tree-sitter>=0.20.0
# tree-sitter-cpp>=0.20.0

